education_admin/requirements.txt
django==3.2
flake8==4.0
python-dotenv
psycopg2-binary==2.9.3
django-split-settings==1.1.0
Faker==13.14.0
django-debug-toolbar==3.2.4
backoff
djangorestframework
Pandas
openpyxl
pydantic
celery
redis
drf-yasg
docutils
